requirements.txt
GitPython
jinja2
